requirements.txt
scikit-learn
tensorflow
pandas
dvc
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure
joblib
types-PyYAML
scipy
Flask
Flask-Cors
kagglehub
-e .